backend/pyproject.toml
[tool.poetry]
name = "wmo-data-ingestor"
version = "0.1.0"
description = "Ingestion pipeline for climweb data"
authors = ["Florent Gravin <florent.gravin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.1"
flask-cors = "^6.0.1"
flask-migrate = "^4.1.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
geoalchemy2 = "^0.17.1"
python-dotenv = "^1.1.0"
flask-httpauth = "^4.8.0"
pytest = "^8.4.1"
pytest-dotenv = "^0.5.2"
pandas = "^2.3.0"
geopandas = "^1.1.1"
pytest-mock = "^3.14.1"
healthcheck = "^1.3.3"
gunicorn = "^23.0.0"
pytz = "^2025.2"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
